[FIX] : vision_transformer model migration
The existing model was created with Pytorch and operated with ONNX
Runtime for Android Serving. Howerver, since it is generally CPU-based
inference, there was inefficiencies in terms of speed and cost,
necessitating model optimization to overcome these problems.

We are considering hardware acceleration, but NNAPI is going to be
deprecated, we decided to use tensorflow lite GPU delegation.
To improve the performance of VisionTransformer, which needs to analyze
tens of thousands of images , we migrated it to the tflite model first.

For the TextTransformer, we are experiencing difficulties with the model
conversion, so it is planned to be introduced later.